Updated Add/Remove functionality and localstorage
