+++ src/index.js
@@ -1,7 +1,8 @@
+/* eslint-disable no-loop-func */
 import './style.css';
 import listContainer from './modules/listContainer.js';
 import handleUserInput from './modules/handleUserInput.js';
-// import tasks from './modules/tasksData.js';
+import checkLocal from './modules/checkLocal.js';
 
 const app = () => {
   const main = document.getElementById('main');
@@ -9,4 +10,5 @@ const app = () => {
 };
 
 app();
+checkLocal();
 handleUserInput();
